config package refactor
