+++ components/cms-controller-manager/pkg/controller/docstopic/config.go
@@ -1,6 +1,7 @@
 package docstopic
 
 import (
+	"github.com/kyma-project/kyma/components/cms-controller-manager/pkg/webhookconfig"
 	"github.com/vrischmann/envconfig"
 	"time"
 )
@@ -8,8 +9,7 @@ import (
 type Config struct {
 	DocsTopicRelistInterval time.Duration `envconfig:"default=5m"`
 	BucketRegion            string        `envconfig:"optional"`
-	WebhookCfgMapName       string        `envconfig:"default=webhook-config-map"`
-	WebhookCfgMapNamespace  string        `envconfig:"default=kyma-system"`
+	webhookconfig.Config
 }
 
 func loadConfig(prefix string) (Config, error) {
