+++ components/cms-controller-manager/pkg/webhookconfig/asset_webhook_config_service.go
@@ -1,4 +1,4 @@
-package config
+package webhookconfig
 
 import (
 	"context"
