+++ components/cms-controller-manager/pkg/webhookconfig/automock/asset_webhook_config_service.go
@@ -2,7 +2,7 @@
 
 package automock
 
-import config "github.com/kyma-project/kyma/components/cms-controller-manager/pkg/config"
+import config "github.com/kyma-project/kyma/components/cms-controller-manager/pkg/webhookconfig"
 import context "context"
 import mock "github.com/stretchr/testify/mock"
