+++ components/cms-controller-manager/pkg/webhookconfig/webhook_service_config.go
@@ -1,0 +1,6 @@
+package webhookconfig
+
+type Config struct {
+	WebhookCfgMapName              string        `envconfig:"default=webhook-config-map"`
+	WebhookCfgMapNamespace         string        `envconfig:"default=kyma-system"`
+}
